Split PromoterService to single-responsible services
