+++ Capri.Services/IPromoterDeleter.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Threading.Tasks;
+using Capri.Database.Entities;
+
+namespace Capri.Services
+{
+    public interface IPromoterDeleter
+    {
+        IServiceResult<Promoter> Delete(Guid id);
+    }
+}
